fix(grin): remove the FC1 lowering bridge
