+++ components/aihc-grin/test/Test/Fixtures/fc2-grin/basic.grin
@@ -9,9 +9,9 @@ caf "Test\NULanswer"%2 :: BoxedRep Lifted = (Fanswer_thunk)
     eval @(BoxedRep Lifted) (x%0 :: BoxedRep Lifted)
   case ($grin_scrutinee_whnf_1%1 :: BoxedRep Lifted) as scrutinee%0 :: BoxedRep Lifted of
     data Test.False ->
-      constant (Test.True%-2974643431 :: BoxedRep Lifted)
+      constant (Test.True%-2469803219 :: BoxedRep Lifted)
     data Test.True ->
-      constant (Test.False%-2526026414 :: BoxedRep Lifted)
+      constant (Test.False%-2486283166 :: BoxedRep Lifted)
 
 answer_thunk -> BoxedRep Lifted =
-  call @(BoxedRep Lifted) "$entry$Test\NULnot" (Test.False%-2526026414 :: BoxedRep Lifted)
+  call @(BoxedRep Lifted) "$entry$Test\NULnot" (Test.False%-2486283166 :: BoxedRep Lifted)
